+++ lib/baby_squeel.rb
@@ -33,11 +33,9 @@ end
 ActiveSupport.on_load :active_record do
   require 'baby_squeel/active_record/base'
   require 'baby_squeel/active_record/query_methods'
-  require 'baby_squeel/active_record/calculations'
   require 'baby_squeel/active_record/where_chain'
 
   ::ActiveRecord::Base.extend BabySqueel::ActiveRecord::Base
   ::ActiveRecord::Relation.prepend BabySqueel::ActiveRecord::QueryMethods
-  ::ActiveRecord::Relation.prepend BabySqueel::ActiveRecord::Calculations
   ::ActiveRecord::QueryMethods::WhereChain.prepend BabySqueel::ActiveRecord::WhereChain
 end