+++ lib/baby_squeel/active_record/base.rb
@@ -5,12 +5,6 @@ module BabySqueel
     module Base
       delegate :joining,
                :selecting,
-               :plucking,
-               :averaging,
-               :counting,
-               :maximizing,
-               :minimizing,
-               :summing,
                to: :all
     end
   end